add vector

+++ ref_str.h
@@ -8,8 +8,8 @@ extern "C" {
 struct ref_str_t;
 typedef struct ref_str_t ref_str_t;
 
-/* 不允许NULL赋值,会自动更新为""  */
-ref_str_t *rs_ini(const char *str, int len);
+/* len=0表示默认长度  */
+ref_str_t *rs_ini(const char *str, size_t len);
 int rs_fini(ref_str_t *rs);
 
 /* 
@@ -25,23 +25,15 @@ ref_str_t *rs_use(ref_str_t *rs);
  */
 ref_str_t *rs_move(ref_str_t *rs);
 
-/* 
- *
- * 添加辅助结构,限定边界
- * 如果你非要超出去,我这里也没有办法啊
- * 需要在use之后才能用,并且在fini之后不能使用
- *
- */
-
 struct ref_str_data_t {
-    int begin;
-    int end;
+    size_t begin;
+    size_t end;
     const char *str;
 };
 typedef struct ref_str_data_t ref_str_data_t;
 
 ref_str_data_t rs_get(ref_str_t *rs);
-int rs_set_range(ref_str_t *rs, int begin, int end);
+int rs_set_range(ref_str_t *rs, size_t begin, size_t end);
 
 #ifdef DEBUG
 void rs_debug(ref_str_t *rs);
